fixing workflow - fixing conftest.py and docker-compose.yml and helpfer functions
